requirements/base.txt
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-coreapi==2.3.0
django-cors-headers==4.0.0
django-cron==0.6.0
django-debug-toolbar==4.1.0
django-extensions==3.2.1
django-filter==23.2
django-health-check==3.17.0
django-jazzmin==2.6.0
django-redis==5.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.14.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
